fix(core): reset agent loop epochs safely
feat(agent): upgrade generation to long and fix race conditions in scheduler lifecycle

Change `IAgentLoopScheduler.Generation` from `int` to `long` to prevent overflow in long-running sessions. Replace lock-based access with `Interlocked.Read` for lock-free reads.

Fix race condition where `Clear()` during an active tick could allow the old epoch to overwrite reset metrics. Add generation check after agent execution to abort stale ticks.

Fix `CompPawnAgent` registration race: if scheduler generation changes during registration, unregister and retry on next tick instead of caching false ownership.

Fix `RimMindRuntimeGameComponent` to eagerly resolve services on each lifecycle event instead of lazy initialization, ensuring proper service replacement.

Add comprehensive tests for generation type, race conditions, lifecycle reset ordering, and service replacement behavior.

+++ Source/Application/Common/Interfaces/Agent/IAgentLoopScheduler.cs
@@ -4,7 +4,7 @@ namespace RimMind.Application.Common.Interfaces.Agent
 {
     public interface IAgentLoopScheduler
     {
-        int Generation { get; }
+        long Generation { get; }
         bool Register(string key, AgentLoopKind kind, IAgentControl agent);
         bool Unregister(string key);
         IAgentControl? Find(string key);
