+++ Tests/Presentation/Agent/AgentScopeExecutableTests.cs
@@ -305,7 +305,7 @@ namespace RimMind.Tests.Presentation.Agent
                 _events = events;
             }
 
-            public int Generation { get; private set; }
+            public long Generation { get; private set; }
             public int RegisterFailuresRemaining { get; set; }
             public Exception? RegisterFailure { get; set; }
             public Dictionary<string, Exception> UnregisterFailures { get; } = new();
